Move Rust code to `model_processing`
Also add tests for parse_arg_for_paths
