+++ model_processing/crates/webify_models/src/cli/create_progress_bar.rs
