+++ model_processing/crates/webify_models/src/cli/parse_args_for_path.rs
@@ -1,0 +1,63 @@
+//! Extract the path from the arguments provided, or return errors when there is none
+//! or if a file is provided
+
+use std::{
+  io::{Error, ErrorKind},
+  path::Path,
+};
+
+pub fn parse_args_for_path<'a>(
+  args: &'a Vec<String>,
+) -> std::result::Result<&'a std::path::Path, std::io::Error> {
+  if args.len() <= 1 {
+    return Err(Error::new(
+      ErrorKind::Other,
+      "Path not provided, no work to do.",
+    ));
+  }
+
+  let path = Path::new(&args[1]);
+  if !path.is_dir() {
+    return Err(Error::new(
+      ErrorKind::Other,
+      "Path provided is a file, please provide a directory.",
+    ));
+  }
+
+  Ok(path)
+}
+
+#[cfg(test)]
+mod tests {
+  use super::*;
+
+  #[test]
+  fn test_correct_arg_is_returned() {
+    let mut args: Vec<String> = Vec::new();
+    args.push(String::from("foo/bar")); // Test base path
+    args.push(String::from("./tests/cli")); // Test provided arg path
+
+    // Return the first argument as a path
+    let arg = parse_args_for_path(&args).unwrap();
+    assert_eq!(arg, Path::new("./tests/cli"));
+  }
+
+  #[test]
+  fn test_error_on_no_path() {
+    let mut args: Vec<String> = Vec::new();
+    args.push(String::from("foo/bar")); // Test base path
+
+    let arg = parse_args_for_path(&args);
+    assert!(arg.is_err());
+  }
+
+  #[test]
+  fn test_error_on_path_to_file() {
+    let mut args: Vec<String> = Vec::new();
+    args.push(String::from("foo/bar")); // Test base path
+    args.push(String::from("tests/README.md"));
+
+    let arg = parse_args_for_path(&args);
+    assert!(arg.is_err());
+  }
+}
