+++ model_processing/crates/webify_models/src/image_processing/convert_to_png.rs
