+++ model_processing/crates/webify_models/src/image_processing/image.rs
