+++ model_processing/crates/webify_models/src/image_processing/move_to_textures_dir.rs
