+++ model_processing/crates/webify_models/src/image_processing/scan_dir_for_images.rs
