+++ model_processing/crates/webify_models/src/mesh_update/scan_dir_for_meshes.rs
